Rename client package itself
